demo/requirements.txt
gradio_bbox_annotator